+++ src/cordova-obj.js
@@ -37,6 +37,8 @@ CordovaClone = function() {
 	 */
 	function parseContents() {
 		that.jsonObject = JSON.parse(this.response);
+		var deviceEvent = new Event('config_updated');
+		document.dispatchEvent(deviceEvent);
 	}
 
 	/**
@@ -92,6 +94,10 @@ CordovaClone = function() {
 	}
 
 	this.readConfig();
+
+	this.setInterval(function() {
+		that.jsonObject = that.readConfig();
+	}, 'updater', 5000);
 
 	CordovaClone.prototype.singleton = this;
 	return this;