+++ src/plugins/device.js
@@ -1,0 +1,6 @@
+/**
+ *  Holds the device object
+ */
+document.addEventListener("config_updated", function() {
+	window.device = cordovaClone.jsonObject.device;
+})