+++ version-history.md
@@ -1,8 +1,10 @@
 Current Version
 ================
-# 0.2
+# 0.2.1
 
 * Simulated navigator.geolocation object with delay features for testing
+* Device object
+* Configuration interval for every 5 seconds
 
 # 0.1